backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-extra-fields==3.7.0
filetype==1.2.0
mysqlclient==2.2.4
pillow==10.3.0
PyMySQL==1.1.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
